Refactor package/directory structure, replace KubeMQ with RabbitMQ
#14
