--- SQIP-Model.drawio
@@ -1,1 +1,0 @@
-<mxfile host="Electron" modified="2020-03-05T16:42:14.272Z" agent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_3) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/12.6.5 Chrome/80.0.3987.86 Electron/8.0.0 Safari/537.36" etag="AK5Tuq-7jOomu0taGiNi" version="12.6.5" type="device" pages="2"><diagram name="SQIP Package Diagram" id="8QqmKJ0gqcGHcXEKBI9b">7VpRd6I4FP41PpYDQdQ+tlo7u9Oe7W5nZ3ef5kSIkDEQGoLV/vq5CUFFEWmndPZBjkfN5eYmud/HdwPac8fx6lbgNLrnAWE9ZAernjvpIeQ49gA+lGVdWEauXRhCQQPjtDU80hdijKVbTgOSVRwl50zStGr0eZIQX1ZsWAj+XHWbc1YdNcUhOTA8+pgdWv+hgYw267K3Jz4RGkZm6JFnTsywvwgFzxMzXg+5c30Up2NcxjL+WYQD/rxjcm967lhwLotvfPZdrQ/ZDM8gxyam6rkkIqM8qcy3zG1OxhStvk7kzRp/mV8sHz7b9GIbPl6NCQPApkUTvphhKmebo20yJEgiT41ajlS/Gk+jgTwTkWGfRAAYrK/94tDB4uxMrks4s2caM5xA63rOE/loztjQ9iPKgju85rlaRSYBv7J1HXFBX8Afq6k6YIDTQhq2ooGKRhkbc8YFGBKuB9h2elTBzDCCZNDtocyWs2e6x6uK4x3OZDlBzhhOMzrTU1YdYyxCmlxzKXlsnDaUkGS1x+AT+Dg7ibslPCZSrKGfidI3NDWX8bCk7fPORVH6RDvXw3BkrkUDYbiJjHZpZ7CqY2EdVb6WlL/qqYl45SXwc9TpN1IHEio18oIvyB7UNehjRkN1TTIyV90UIhRk5cqYJU9VsBT7NAnvtM+kv7X8ZTKoTBz6zpnWhogGAUkUObjEEs825E05TaReqncNL8jJ2La8ngcTH0Pb2bbhpdyFHPME1oKpJgUBmj0TRbWfYBCqZ5ChDBq0Y0zp93bGtNMtQ5UlZrlBeIIlKShluxDSBo8psjcCfabCe1HBQ51RobvCMmijDiXUPqSHiBqwY4BN9WmHdw2ZOoRbV8SipjmjvQrpHPAdNjS2PvSkYJ9DAuPXVQ1yLy8tx7vcHsNKTeofliTX/iCBuXdeJunnJ8xknt6Of//2NL8qykmFJCSAvaVpAg4RD3mC2c3WupfIrc8dVyKhjd+JlGsDE84lVzsNGZcbE8ivWP9ruKEb/+02JqtKS2XORWXzgQgKmVC81T77fCjWoxbxBlQhETwXPmlIWLm/h10NkQ1+R8RFEIYlXVYn9wqQ26HqnVHtBlW3G1T/zoj4o6Yo4DQtYjGaLApTJKW6q7xSodA0pDLKZ5YPW2s0HdMl9a98CZvMDJqPf/72AB9SEKKGAwmF9KLpJmYjg3ZFruYOJcKp+jrX9alOhYsy8aWgjkIUykF5b9o37c0tqdMvDA88o7K4SzTbkB16Vfl5XOZrylE9m7rQfm9P7A/Vflgj9gP7LWK/Q5pXSMPlWRpeJw2opTSMfqXgO2fF7whWp98NrsckPxV8SfWmvxPhn+7Fb2TVicdU5yJQVwS8y77lDOzN4VRKglv3TMqpqQneB9YEt7f/lCGXlB0oyhnwNwDeH3UHeDt0h/+j0gA3+ceLw6b1q2vDoG1t+NjSMGc4WzSXBe1ipZgxIjOQehUkM1WCJPDmWI610r4n1b/5WdJZDGrFYP83iI+9BTi6qVg/sWDWTJ10nVKLi7DYIxQxpmXHNnwZnfnyer6M7CpfUM3+oI4v77M9OMaXRT4j8VMzXwLuZ1bhaFGuDGRJGE9jlQ7FHKgZiaXRQK5+GHzB5xc+5FU7nH4C0fxL15lPrfjUVn/eh09g2v4ZQLvv/OPCvfkB</diagram><diagram name="DIAGRAM TEMPLATE" id="9f46799a-70d6-7492-0946-bef42562c5a5">7VbbbtswDP0aP27wJW7Sx8a9YWiHYRm2Z8VWbK2yaMhKk/TrR8lUbSdZF2zN2wwjsY5IiuI5ugRJVm/vNGuqRyi4DOKw2AbJdRDHUXQxwT+L7DpkNiWg1KIgox5YiBdOYEjoWhS8HRkaAGlEMwZzUIrnZoQxrWEzNluBHI/asJIfAIucyUP0hyhM5ecVhn3HPRdlRUPPUupYsvyp1LBWNF4QJyv3dN0187HIvq1YAZsBlNwESaYBTPcFy592fnEo2RJrTDGt5zPXrQDVQdsO87XtI9XbjEvk5rZr4gdFHPWGYr88miszTPT2rXRSV844JWfJcl5hxTHBUVQ/yPpb9JwW9zWLdfspvFmIaf75Q3yQctianeej3YhaMoWt+QqUWVBPiO28ErJ4YDtY24RbgwT41rwCLV7QntlUIwSwWxuSW3xhowkpM5CgEVDgBuidFjYYDaN5i25ffGGiPeiRbUeGD6w1PkGQkjWtWLqUrWPNdCnUHIyBmoxeOTV8+1sqokGN7jjU3OgdmpDDhCRFS27qJbYZCNjbVAPtTme0boit8jVyPNQN0XJMRsdU8d3L8yqwiaRerv+mksmbKsHaGUeyhie+x+oRopkUpV0/kq+smy2+wC3gimADjQ3WsFyo8sHZXE965CtV0EKAvivp1nElioIrqwMwzLDlq04bEMq4qaZzfLEmWfgxDVJMPMN21LfxtebaZKBwLkw4/jkqasOtqk4Tyx9X2UBBJBlcEScpxtv9vWIONp6j2ZJUnplcE8PXzPBOUnWNIYsCf3b40Gb6XwrvJYU0PpsUzneGXJyyO3iqcywP10fIrpE263Ma30fEdEa63eHXHV/RbO8wjA70jpeP0D0uKbyT8ILs3uG4SS4vP0bpZf9MR8fP5PD0ScKz7CXY7O9LznRwKU1ufgE=</diagram></mxfile>
