+++ architecture/SQIP-Model.drawio
@@ -1,0 +1,1 @@
+<mxfile host="Electron" modified="2020-03-12T17:15:27.363Z" agent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_3) AppleWebKit/537.36 (KHTML, like Gecko) draw.io/12.6.5 Chrome/80.0.3987.86 Electron/8.0.0 Safari/537.36" etag="Nh-GrTF7DwyDMjnDdDYE" version="12.6.5" type="device" pages="3"><diagram name="SQIP Package Diagram" id="8QqmKJ0gqcGHcXEKBI9b">7VrZcts2FP0aPUbDRdTyKFOW28aeqnWaNk8ZiIRIxCDBgNDmr+8FCG4STVOOpeRBHM1YuLjY7jk4F6Dcs91od8dREj4wH9OeZfi7nj3rWZZpGkP4Iy17bTEMO7MEnPjaVhoeyTPOHbV1TXyc1hwFY1SQpG70WBxjT9RsiHO2rbutGK2PmqAAHxkePUSPrf8SX4TFwoyy4jdMglAPPXZ0xRJ5TwFn61iP17PslXqy6gjlfWn/NEQ+21ZM9m3PdjljIvvGlt/k+iyDoiUEWfcpW24wTwmLM5O25bFdY5dYu88zcbtHn1YfNouPBvlQdh/tXEwBsXlWhC96mFpte29FhDiOxWuj5iM1r8ZRaFiO7pEiD4cAGKyvBkbrdKyjxRmp2OdwplsSURRD6WbFYvGoawwoeyGh/j3as7VcRSoAv7x0EzJOnsEfyamaYIBqLjRbraHsjVDqMso4GGKmBigbPcrO9DAcp9BskUfLPDA9oF3N8R6lIp8goxQlKVmqKcuGEeIBiW+YECzSTh3BMAvqCLyrBFcH7g6zCAu+BxddO9A01ft4lNN2W9kUuU9Y2Q+jsd6LGsKg6Nmq0k5j1cTCJqp8zik/7cmJOPkW+DHqDFqpA4ESCnnOnvAB1A3oI0oCuScpXslmMtIEZGWqzYIlsrMEeSQO7pXPbFBa/tYRlCYGbVdUaUNIfB/HkhxMIIGWBXkTRmKhlurcwAdi4hp9p+fAxF0om2UZPtKdC5fFsBZEFE8w0GyLJdW6MsjqzCBNGWvYjTG539sZ0023NFU2iK41wjMkcEYpmaTmphzPMgqBvlLhvajgWGejwvkSy7CLOuRQexAxzBvAjgA22aYb3g1kOiPcKiNmOc0cH2RI84jvcKAx1KMmBecc7Gu/c+UgezLpm86kfEa1nDQ4Tkm2cSGBeTCfZ8nH74iKdXLn/vH1+2qapZMaSbAPZ0tdBBxCFrAY0dvSehDI0ueeSZFQxm9YiL2GCa0FkycNEeUHEwg53/+nuaEKX6qF2a5WkpGzrby4wJxAJCRvlc8hH07DNWVr7uGW+OTHeTjEYNHip8VCBq+VJRxTJMimfnA/AeRuqDpXVN8FVftCqP6TYv5nQ1JASZL1RUn8lJlCIeStciq7suYBEeF62ffgaG3NXbIh3tQTcMhMofj41+8L+CM4xnI4kFAIrzUv+mxlUHWhDTeUECXy60rlpyYVztLEp4w6ElFIB/nddKDLxZXUHGSGBUuJyG6J+hhSoVedny/LfEM6+jE2naL9zoHYH6v9qEHsh8ZbxL5CmhOkYXKVhlZpsDpKw/iXEnzzqvjvA6s5+MmSn3C2IerQfxbhnx/038qqV15TXZNAUxJwJoO+OTSKx6ylBLvpnZTZkBOcC+YEfcypvGVYC0KPFOUK+BsAH4zPB3g3dEe/UGqAS/7LyaEoXTg3DLvmhp+cGlYUpU/taUG59BNEKRYpSL3sJNVZAsfyfWHf7O+U76vq3/4u6SoGjWJw+BvEZa8ALzGHSs7yDpiPr5ifjvnYqGNuNeX4ppd875PjXwI9wmmKAoCjA+5m+89KV+A7Ad+02c8IPJjKn96Ve+UfHOzb/wE=</diagram><diagram name="SQIP Class Diagram" id="Eco0ieIgOOWpat-biCDg">7VbbbuIwEP2aPLbKhbTwWNLbSq32grT7bBKTeOt4Iscs0K/fsRkTAlGL2uVtIyTi4/GMZ86xJ0GS1esHzZrqGQougzgs1kFyG8RxFIVX+GeRDSFhmGyRUouCsA6YiVfuDQldioK3PUMDII1o+mAOSvHc9DCmNaz6ZguQ/agNK/kRMMuZPEZ/icJUu8TCbuKRi7Ki0OOUJuYsfyk1LBXFC+Jk4Z7tdM28L7JvK1bAag9K7oIk0wBm+wbz3za/OJRsjkUmn3blH65bAaq3X1/bL7P0azkZP8z1BX+8eI6+V3l+0bmv1xmXyNj9dogvFKY3+7a3XYU0V+a9qD7ScDapYyNOyaNkOa+QMMzv9OTio+TC1mw8ne1K1JIpHE0XoMyMZkIc55WQxRPbwNJm0Rrkz4+mFWjxivbMbjVCAKe1IbXGV9abkDIDCRoBBS5At2hmnVEYzVtc9s1XKzqAntm6Z/jEWuM3CFKyphVzt2W7sGa6FGoKxkBNRjtJGL4+UPA7/ER7hXvgUHOjN7iOvIxIpnSOr71sV3uHwttUe+fhekxnkSgsd57jfdkRV0MqHJLKTy/5m8BuJPVH4HPSSd6UDhbUOOY1vPADqgfYZ1KU9kxKvrDLLCMCr5Ubgg001lnDcqHKJ2dzO+qQH1RBCwGuXUh3N1SiKLiy4gDDDJvvxNuAUMalmk7xhzXJwss0SHHjGY6jbow/a65NBgpzYcKJgqPMVtxK7RMKiocVRJKJr05TjLf7uGJOu7dGlCiTS2L4lhm+lZRtUveRjReHuwv6vxT+lRTS+GxSOF9jSU+5HTzVOZaH6wGya6TNrjmN7wExnZFu1xG3PS0aH3TI6Ejv+EETusdtCr9zeEF25+pByWRyGaWT7rnu9aTRcUtKwrNcMDjsPsyc6d7nb3L3Fw==</diagram><diagram name="DIAGRAM TEMPLATE" id="9f46799a-70d6-7492-0946-bef42562c5a5">7VbbjpswEP0aHrviEpLs44bsRdVuVTVV++yAA+4aBhnntl/fMYwDBLqK2s1bEUrw8cx4POf44gRRfnhUrMxeIOHS8d3k4ARLx/c9z53in0GOhLhu0CCpEglhLbASb9waEroVCa96hhpAalH2wRiKgse6hzGlYN8324Dsj1qylA+AVczkEP0pEp2dJua2HU9cpBkNPQ+pY83i11TBtqDxHD/Y1E/TnTMbi+yrjCWw70DBvRNECkA3X7D+Zebnu5KtscgU03juuKoEFA10aDBb2zZSfoi4RHIemiZ+UMReryvOy6N4obuJPryXTliX0w/JWbKYZ1hxTLAX1Q6y/e7twuQpZ76qPrv3KzGLv3zyBym7lT5aPqq9yCUrsLXYQKFX1ONiO86ETJ7ZEbYm4UojAba1yECJN7RnJlUPAexWmuTmT000IWUEEhQCBdQDtE4rE4yGUbxCt6+2MN4Z9MIOPcNnVmmbIEjJykqs65SNY85UKooFaA05GQ3r7p1o1vzQgahGjxxyrtURTah3QpKiNTezEtt3BGxtso52Z3NaN8RWeorsd3VDtIzJaEwVP6w87xyTSGjl+m8qmbyrEiyUrklW8MrPWB0hmkmRmvUj+ca4mUoL3ALuCNZQmmAli0WRPtc2y0mLfKMKGgjQdyPrdZyJJOGF0QFoptn6pNMSRKHrqYYLfLEmkXsTOiEmHmHba9v4GnOlIyhwLkzUkuCoqD03qhoRy5+X1EUKIsn408sUY+3+XjGDjWd0AiSVHZNbYnjJNG8klecYMknw54gPbab/pfBRUgj9q0nhemfI9JLdwVIdY8W4GiE7R9qMz2V8j4jpinTXh19zfHnzs8PQG+gdLx9u/dRJ4Z2EJ2T3AcdNcHt744W37TPrHT+T4ekTuFfZS7DZ3pdq086tNLj/DQ==</diagram></mxfile>
